feat: Add AppLogger class for enhanced logging functionality
